Adds support to composer installed AWS SDK
In our Magento projects we use firegento/psr0autoloader to support autoloading of packages installed by Composer. We would like to use newer versions of AWS SDK by installing it with Composer in our project. But, in such situation, the require of `AWSSDKforPHP/aws.phar` fails because that file only exists in the PEAR package. With this change, the require of `AWSSDKforPHP/aws.phar` is done only if the `Aws\Sqs\SqsClient` class is not already loaded.